+++ src/main/java/karlord19/cardarchitect/Position.java
@@ -1,11 +1,11 @@
 package karlord19.cardarchitect;
 
 public class Position {
-    public int x, y;
+    public float x, y;
     public String toString() {
         return "pos: (" + x + ", " + y + ")";
     }
-    public Position(int x, int y) {
+    public Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
